pull out custom pocketbase logic into their own files
